Leaderboard April 3 release (#309)
This PR is for the leaderboard April 3 release:

1. Bug fix for evaluation dataset possible answers, including those that
are identified in #301.
2. Implement **string standardization** for the AST evaluation pipeline,
i.e. removing white spaces and a subset of punctuations (`,./-_*^`) to
make the AST evaluation more robust and accurate.
3. Fix AST evaluation issue for type `tuple`. 
4. Fix AST evaluation issue for Java and JavaScript.
5. Add 2 new models `meetkai/functionary-small-v2.4 (FC)`,
`meetkai/functionary-medium-v2.4 (FC)` to the leaderboard.

This PR **DOES** change the leaderboard score. We will update the
leaderboard website shortly, in a different PR.

---------

Co-authored-by: Charlie Cheng-Jie Ji <charliechengjieji@berkeley.edu>
Co-authored-by: Fanjia Yan <fanjiayan@berkeley.edu>

---------

Co-authored-by: CharlieJCJ <charliechengjieji@berkeley.edu>