BFCL April 19th Release (Dataset & Pipeline) (#377)
This PR is for the BFCL April 19th Release. In this release:


- Bug fix for the evaluation dataset in the executable test categories.
This includes updates to both prompts and function docs.
- The `evaluation_result` field has been removed to accommodate the
variability in API execution results across different evaluation runs.
Instead, a human-verified `ground_truth` is now included for the
executable test categories. During each evaluation run,
`evaluation_result` is generated anew using the `ground_truth`, and then
compared against the model output.
- A stricter metric has been adopted when using the `structural_match`
(aka. type match) evaluation criteria ---- For `list` results, the
lengths are compared; for `dict` results, the keys are matched. This is
to account for the fast-changing nature of some of the real-time API
results while ensuring the evaluation remains meaningful.
- Added another evaluation criterion `real_time_match` for the
executable category, which is a looser form of `exact_match`
specifically for numerical execution results. The execution result must
be within a certain percentage threshold (20%) from the expected result
to accommodate the live updates of API responses. Users can change this
threshold value in `eval_checker_constant.py`.
- Added support to distinguish Cohere's optimized score vs. original
score.
- Resolved #363 

This PR **DOES** change the leaderboard score. We will update the
leaderboard shortly, in a different PR.
We will also update our HuggingFace dataset accordingly.

---------

Co-authored-by: Charlie Cheng-Jie Ji <charliechengjieji@berkeley.edu>
Co-authored-by: Fanjia Yan <fanjiayan@berkeley.edu>