Small corrections to possible_answers for simple test category (#405)
I’m redoing the leaderboard results and scores for some models in the
`simple` test category. I’ve noticed the model is marked down for
answers that seem correct to me. Here are my suggestions for the
corrections.

This PR DOES change the leaderboard scores, which will be updated later.