[BFCL] Dataset Question Fix for Executable Parallel Category (#568)
This PR fixes #541.

Thanks to @XuHwang for pointing this out. Some prompts in the
`executable_parallel_function` category are misleading. They are
formatted as:
```
First let's start by xxxx. Once we have that xxxx, then xxxx. 
One at a time?
```
This would suggest the model generates one function calling at once,
resulting in multi-turn functional calling. So we rephrased the prompts
to make it clear.

The possible answers are unchanged. 

We will update the leaderboard shortly, in a different PR.