[BFCL] Use `N/A` in Score Report for Unevaluated Categories (#849)
Previously, when a test category was not evaluated, it was represented
as a “0” in the score CSV files. This made it difficult to distinguish
between a category that was unevaluated and one where the model actually
scored a 0. In this PR, unevaluated categories will be marked as “N/A”
to clarify this distinction.
Additionally, this PR refactors the code related to the score report
output section to reduce duplication and improve maintainability.

**Note**: This change will not affect leaderboard scores. If a category
is unevaluated, it will still be treated as a 0 when calculating the
overall accuracy, and the overall accuracy column will report the score
with 0 taken into account.
For summary columns that are the average of a few categories (for
example, python simple (which consists of `simple`, `java`, and
`javascript`), if any of the categories involved are unevaluated, the
summary column will be marked as 'N/A` to avoid confusion.